requirements.txt
flask
tensorflow
numpy
pillow
werkzeug
streamlit